% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMGPEffect.R
\name{plotMGPEffect}
\alias{plotMGPEffect}
\title{Human MGP}
\usage{
plotMGPEffect(obj, comp, type, sdy, lambda, ncores)
}
\arguments{
\item{obj}{runHumanMGP object}

\item{comp}{Which PLS component to visualize (default = 1)}

\item{type}{Type of plot (options: "max", "min", "overlay", "heatmap"; default = "none")}

\item{sdy}{Standard deviation of the yscores. sdy=3 will show the effects of -3sd vs +3sd (default = 3)}

\item{lambda}{Regularization parameter of the TPS (only valid for sparse landmarks; default = 1e-8)}

\item{ncores}{Number of CPU cores to be used for parallellization (default = 1)}
}
\value{
\item{predmax }{Predicted MGP effect at +sdy standard deviations from the mean}
\item{predmin }{Predicted MGP effect at -sdy standard deviations from the mean}
}
\description{
Generates associated MGP effects from "runHumanMGP" output
}
\author{
Hanne Hoskens
}
