% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runHumanMGP.R
\name{runHumanMGP}
\alias{runHumanMGP}
\title{Human MGP}
\usage{
runHumanMGP(
  GOterm,
  cohort,
  lm,
  scale,
  covs,
  window,
  ncomp,
  npc,
  signif,
  nperm,
  ncores
)
}
\arguments{
\item{GOterm}{GO term (name or ID) or gene list (names or ensembl ID) of interest}

\item{cohort}{Cohort to be analyzed ("3DFN" or "TANZ")}

\item{lm}{Sparse or dense landmarking scheme to be used ("sparse" or "dense"; default = sparse)}

\item{scale}{Logical value indicating whether or not to scale to unit size (default = T)}

\item{covs}{List of covariates to standardize for (options: "none", "age", "age2", "age3", "sex", "height", "weight", "csize")}

\item{window}{Search window (in kb) around transcription start and end sites (default = 0kb)}

\item{ncomp}{Number of PLS components to consider (default = 1)}

\item{npc}{Number of PCs used to build gene composite score (1 or "all"; default = 1)}

\item{signif}{Logical value indicating whether or not to perform significance testing (default = F)}

\item{nperm}{Number of permutations for significance testing (default = 99)}

\item{ncores}{Number of CPU cores to be used for parallellization (default = 1)}
}
\value{
\item{GOterm }{Search term}
\item{cohort }{Cohort}
\item{PLS }{Summary of pls2B output}
\item{GeneLoadings }{Matrix (nxncomp) containing the relative loadings per gene composite score}
\item{ShapeEffectMax }{Matrix (3xnLMxncomp) containing the shape effects associated with each PLS component}
\item{ShapeEffectMin }{Matrix (3xnLMxncomp) containing the "inverse" shape effects associated with each PLS component}
\item{ShapeAverage }{Matrix (3xnLM) containing the mean shape of the specified cohort}
\item{SV }{Singular values (covariance) per pair of PLS axes; plus significance}
\item{R2 }{Shape variance explained per PLS component; plus significance}
\item{PD }{Procrustes Distance between mean shape and MGP effect; plus significance}
}
\description{
Performs a two-block PLS on sets of genetic and shape variables based on "pls2B" from the Morpho package.
The goal is to find pairs of latent variables that show maximal covariation with a set of genetic markers.
}
\author{
Hanne Hoskens
}
